Tests/LibWeb/Text/expected/HTML/document-named-shenanigans.txt

Testing <div> element
,,,,,,,,,,,,,,,

Testing <form> element
,,,,,,[object HTMLFormElement],,[object HTMLFormElement],,,,,,,

Testing <img> element
,,,,,,[object HTMLImageElement],,[object HTMLImageElement],,,,,,,

Testing <embed> element
,,,,,,[object HTMLEmbedElement],,[object HTMLEmbedElement],,,,,,,

Testing <object> element
,,,,,,[object HTMLObjectElement],,[object HTMLObjectElement],,[object HTMLObjectElement],[object HTMLObjectElement],,[object HTMLObjectElement],,
